+++ pkgs/native_assets_cli/test/build_config_test.dart
@@ -40,7 +40,6 @@ void main() async {
       ..setupHookConfig(
         packageName: packageName,
         packageRoot: packageRootUri,
-        targetOS: OS.android,
         buildAssetTypes: ['my-asset-type'],
       )
       ..setupBuildConfig(
@@ -73,7 +72,6 @@ void main() async {
       'package_name': packageName,
       'package_root': packageRootUri.toFilePath(),
       'supported_asset_types': ['my-asset-type'],
-      'target_os': 'android',
       'version': latestVersion.toString(),
     };
 
@@ -85,7 +83,6 @@ void main() async {
 
     expect(config.packageName, packageName);
     expect(config.packageRoot, packageRootUri);
-    expect(config.targetOS, OS.android);
     expect(config.buildAssetTypes, ['my-asset-type']);
 
     expect(config.linkingEnabled, false);
@@ -98,7 +95,6 @@ void main() async {
       ..setupHookConfig(
         packageName: packageName,
         packageRoot: packageRootUri,
-        targetOS: OS.android,
         buildAssetTypes: ['my-asset-type'],
       )
       ..setupBuildConfig(
@@ -121,7 +117,6 @@ void main() async {
       'package_name': packageName,
       'package_root': packageRootUri.toFilePath(),
       'supported_asset_types': ['my-asset-type'],
-      'target_os': 'android',
       'version': latestVersion.toString(),
     };
 
@@ -133,7 +128,6 @@ void main() async {
 
     expect(config.packageName, packageName);
     expect(config.packageRoot, packageRootUri);
-    expect(config.targetOS, OS.android);
     expect(config.buildAssetTypes, ['my-asset-type']);
 
     expect(config.linkingEnabled, true);