+++ pkgs/native_assets_cli/test/code_assets/config_test.dart
@@ -91,7 +91,6 @@ void main() async {
       ..setupHookConfig(
         packageName: packageName,
         packageRoot: packageRootUri,
-        targetOS: OS.android,
         buildAssetTypes: [CodeAsset.type],
       )
       ..setupBuildConfig(
@@ -103,6 +102,7 @@ void main() async {
         outputDirectoryShared: outputDirectoryShared,
       )
       ..setupCodeConfig(
+        targetOS: OS.android,
         targetArchitecture: null, // not available in dry run
         cCompilerConfig: null, // not available in dry run
         linkModePreference: LinkModePreference.preferStatic,
@@ -116,7 +116,6 @@ void main() async {
       ..setupHookConfig(
         packageName: packageName,
         packageRoot: packageRootUri,
-        targetOS: OS.android,
         buildAssetTypes: [CodeAsset.type],
       )
       ..setupBuildConfig(
@@ -128,6 +127,7 @@ void main() async {
         outputDirectoryShared: outputDirectoryShared,
       )
       ..setupCodeConfig(
+        targetOS: OS.android,
         targetArchitecture: Architecture.arm64,
         targetAndroidNdkApi: 30,
         linkModePreference: LinkModePreference.preferStatic,
@@ -148,7 +148,6 @@ void main() async {
       ..setupHookConfig(
         packageName: packageName,
         packageRoot: packageRootUri,
-        targetOS: OS.android,
         buildAssetTypes: [CodeAsset.type],
       )
       ..setupLinkConfig(assets: assets)
@@ -158,6 +157,7 @@ void main() async {
         recordedUsesFile: null,
       )
       ..setupCodeConfig(
+        targetOS: OS.android,
         targetArchitecture: Architecture.arm64,
         targetAndroidNdkApi: 30,
         linkModePreference: LinkModePreference.preferStatic,