+++ pkgs/native_assets_cli/test/link_config_test.dart
@@ -33,7 +33,6 @@ void main() async {
       ..setupHookConfig(
         packageName: packageName,
         packageRoot: packageRootUri,
-        targetOS: OS.android,
         buildAssetTypes: ['asset-type-1', 'asset-type-2'],
       )
       ..setupLinkConfig(assets: assets)
@@ -53,7 +52,6 @@ void main() async {
       'package_name': packageName,
       'package_root': packageRootUri.toFilePath(),
       'supported_asset_types': ['asset-type-1', 'asset-type-2'],
-      'target_os': 'android',
       'version': latestVersion.toString(),
     };
     expect(config.json, expectedConfigJson);
@@ -64,7 +62,6 @@ void main() async {
 
     expect(config.packageName, packageName);
     expect(config.packageRoot, packageRootUri);
-    expect(config.targetOS, OS.android);
     expect(config.buildAssetTypes, ['asset-type-1', 'asset-type-2']);
     expect(config.encodedAssets, assets);
   });