+++ pkgs/native_assets_cli/test/validation_test.dart
@@ -34,7 +34,6 @@ void main() {
       ..setupHookConfig(
         packageName: packageName,
         packageRoot: tempUri,
-        targetOS: OS.iOS,
         buildAssetTypes: ['my-asset-type'],
       )
       ..setupBuildConfig(